Backend/requirements.txt
openai
tiktoken
google-generativeai
streamlit
python-dotenv
anthropic
curl_cffi
nltk
fastapi
uvicorn
pydantic
python-multipart
fastapi-mail
aiosmtplib
pydantic-settings
aiohttp
sqlalchemy 
passlib[bcrypt] 
python-jose 
pydantic[email]